+++ frameworks/kits/js/common/image_packer_napi.cpp
@@ -149,7 +149,7 @@ STATIC_EXEC_FUNC(Packing)
         context->rImagePacker->AddImage(*(context->rImageSource));
         context->rImagePacker->FinalizePacking(packedSize);
         HiLog::Debug(LABEL, "packedSize=%{public}lld.",
-            static_cast<int64_t>(packedSize));
+            static_cast<long long>(packedSize));
 
         if (packedSize > 0 && (static_cast<uint64_t>(packedSize) < bufferSize)) {
             context->packedSize = packedSize;
@@ -215,7 +215,7 @@ STATIC_EXEC_FUNC(PackingFromPixelMap)
         context->rImagePacker->AddImage(*(context->rPixelMap));
         context->rImagePacker->FinalizePacking(packedSize);
         HiLog::Debug(LABEL, "packedSize=%{public}u.",
-            static_cast<int64_t>(packedSize));
+            static_cast<long long>(packedSize));
 
         if (packedSize > 0 && (static_cast<uint64_t>(packedSize) < bufferSize)) {
             context->packedSize = packedSize;
